backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
autopep8==1.5.7
billiard==4.2.0
boto3==1.21.13
botocore==1.24.13
Brotli==1.0.9
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==5.0.3
django-allauth==0.61.1
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-countries==7.2.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.2
django-extensions==3.2.3
django-nine==0.2.5
django-rest-auth==0.9.5
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
idna==3.6
jmespath==0.10.0
kombu==5.3.5
netaddr==1.2.1
oauthlib==3.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.59
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing-extensions==3.10.0.2
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.0.0
wrapt==1.16.0
yookassa==3.0.1
